+++ src/GuardClauses/GuardValidationGuidExtensions.cs
@@ -14,7 +14,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is empty</exception>
         public static void IfEmpty(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             Guid value,
             [CallerArgumentExpression("value")] string parameterName = null
         )
@@ -34,7 +34,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is equals to the other value</exception>
         public static void IfEquals(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             Guid? value,
             Guid? otherValue,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -56,7 +56,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is different to the other value</exception>
         public static void IfDifferent(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             Guid? value,
             Guid? otherValue,
             [CallerArgumentExpression("value")] string parameterName = null