+++ src/GuardClauses/GuardValidationULongExtensions.cs
@@ -14,7 +14,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when the value is greater than</exception>
         public static void IfGreaterThan(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong value,
             ulong max,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -35,7 +35,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when the value is greater than</exception>
         public static void IfGreaterThan(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong? value,
             ulong max,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -58,7 +58,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is less than</exception>
         public static void IfLessThan(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong value,
             ulong min,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -79,7 +79,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is less than</exception>
         public static void IfLessThan(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong? value,
             ulong min,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -100,7 +100,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is equals to the other value</exception>
         public static void IfEquals(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong? value,
             ulong otherValue,
             [CallerArgumentExpression("value")] string parameterName = null
@@ -122,7 +122,7 @@ namespace PowerUtils.Validations.GuardClauses
         /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
         /// <exception cref="PropertyException">Exception thrown when value is different to the other value</exception>
         public static void IfDifferent(
-            this IGuardClause _,
+            this IGuardValidationClause _,
             ulong? value,
             ulong otherValue,
             [CallerArgumentExpression("value")] string parameterName = null